LC-reverse-integer [Runtime: 1 ms; Memory: 42.6 MB]

+++ math/reverse-integer.java
@@ -1,13 +1,13 @@
 class Solution {
     public int reverse(int x) {
-        int res = 0;
+        long res = 0;
         int lastdigit = 0;
         while(x!=0){
             lastdigit = x%10;
             res = (res*10)+lastdigit;
             x = x/10;
         }
-        if(res < Integer.MIN_VALUE || res > Integer.MAX_VALUE){
+         if (res < Integer.MIN_VALUE || res > Integer.MAX_VALUE) {
             return 0;
         }
         return (int) res;
